refactor(worker): organize agent worker sources
