+++ apps/agent-worker/package.json
@@ -9,11 +9,11 @@
     "predeploy:worker": "pnpm --filter @minpeter/pss-runtime build",
     "predev:worker": "pnpm --filter @minpeter/pss-runtime build",
     "predry-run:worker": "pnpm --filter @minpeter/pss-runtime build",
-    "start": "tsx --conditions=@minpeter/pss-source src/worker-simulation.ts",
-    "start:cli": "tsx --conditions=@minpeter/pss-source src/index.ts",
-    "start:edge-cases": "tsx --conditions=@minpeter/pss-source src/worker-edge-cases.ts",
-    "start:worker-sim": "tsx --conditions=@minpeter/pss-source src/worker-simulation.ts",
-    "test": "vitest run src/request-schema.test.ts src/stress-scenarios.test.ts src/stress-guard-scenarios.test.ts src/cloudflare-durable.test.ts",
+    "start": "tsx --conditions=@minpeter/pss-source src/simulation/index.ts",
+    "start:cli": "tsx --conditions=@minpeter/pss-source src/cli/index.ts",
+    "start:edge-cases": "tsx --conditions=@minpeter/pss-source src/simulation/edge-cases.ts",
+    "start:worker-sim": "tsx --conditions=@minpeter/pss-source src/simulation/index.ts",
+    "test": "vitest run src/request/schema.test.ts src/scenarios/index.test.ts src/scenarios/guards.test.ts src/cloudflare/durable.test.ts",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "typecheck:worker": "tsc -p tsconfig.worker.json --noEmit"
   },
