+++ apps/agent-worker/src/cloudflare/edge-alarm-checks.ts
@@ -5,10 +5,10 @@ import {
   listScheduledCloudflareRuns,
   listScheduledCloudflareSessionPrompts,
 } from "@minpeter/pss-runtime/cloudflare";
+import { AgentDurableObject } from "../worker";
+import { workerStorePrefix } from "../worker/constants";
 import { createQueuedRun } from "./edge-case-helpers";
 import { FailingRunLookupStorage } from "./failing-run-lookup-storage";
-import { AgentDurableObject } from "./worker";
-import { workerStorePrefix } from "./worker-constants";
 
 const sessionKey = "room:demo:user:edge";
 const unclaimableAgent = { resume: () => Promise.resolve(null) };
