+++ apps/agent-worker/src/cloudflare/edge-case-helpers.ts
