+++ apps/agent-worker/src/scenarios/plugin.ts
